Comfy (#146)
* onboard logguru

* model manager wrapper

* working poc

* Expand proof of concept:

- Custom model loader loading from custom directory
- Simple pipeline library and loader.
- Cleanup up default pipeline.
- Dot notation parameter settings.

* Lint fixes

* Expand comfy proof of concept. Image saving node.

* MM

* ported MM for model downloads

* pip

* pip

* wip

* fix?

* req

* hacky but works!

* rm mm

* workingMM

* clip

* added_csam_filter

* still_bnroken

* fix: environment build oddities, enable safety_checker (#159)

* fix: environment build oddities, enable safety_checker

* style: style.py fixes

* merge_ruff

* censor

* working safety checker

* PP work

* switch_to_generic_generator

* is_image_nsfw

* post-processing

* POST_PROCESSORS_HORDELIB_MODELS

* unnecessary imports

* removed img_source and img_mask

* removed_old_stuff

* inpainting refactoring

* fixes

* fix_for_retaining_transparency

* facefixers

* added catch for None image

* prevents ignoring facefixer PPs

* index

* stip background

* somwe_stylefix

* fix

* switch to pypi hordelib

* ensure correct torch installed

* fix: remove model dir env var override

* requirements

* switch alchemist to comfy

* ensure_nataili_uninstalled

* later cuda

* fix: cudatoolkit to 11.8 (#161)

* fix: Change cudatoolkit to 11.8

* Merge remote-tracking branch 'upstream/main' into comfy

* Revert "Merge remote-tracking branch 'upstream/main' into comfy"

This reverts commit 3a3bcd18b9f76194a0edde635c04f5de526ce6f3.

* version_bump

* feat: adds hordelib version to top right of terminal ui

* fix: remove obsolete config settings

* fix: load safety checker on CPU only (#165)

* re-enable esrgan 2x

* fix: don't modify a list we're iterating over in another thread

* stye: Bring up to date with ruff/black (#167)

* fix: avoid unloading dynamic models in queued jobs

* fix: default to hordelib stable config

* feat: now attempts to preload all controlnet annotators (#168)

* feat: now attempts to preload all controlnet annotators

* style: lint fix

* Ignore models_to_skip for TOP n models (#169)

* use round() instead of int() for parsed RAM values (#160)


* fix: try not to send jobs for models we unloaded

* feat: adds benchmark command

* fix: important version bump for hordelib

* fix: allow maintenance mode for workers with short names

* Change bridgeData to new domain name (#171)

Also added a bit about the models_to_skip being ignored for TOP n models

* feat: add new resource usage settings to default config template

* Revert to cuda 11.7 as linux has trouble finding the libraries with 11.8 (#174)

Added benchmark for linux

* back_to_118

* bad link

* Bump hordelib version

* fix: don't suggest running update-runtime as admin.

* fix: make the terminal ui a bit more robust

* fix: add pix2pix to models to skip

* feat: support new hordelib settings

* fix: pin hordelib

* Adds scribe worker

* Unpin hordelib

* fix: Import TerminalUI only when used

* avoid terminal ui on scribe

* fix: prevent KAI Client down crashing worker

* fix: terminal ui should report compvis models

* fix: change kudos per hour to be an actual measurement

* fix: respect stats output config time

* fix: kudos displays 0 after 1 hour

* feat: add (r)eset stats option to ui

* feat: adds 5 min job pop time to ui, lint fixes

* fix: defers cn annotator check until after config is loaded (#176)

* style: lint/formatting

* fix: defers cn annotator check until after config is loaded

* style: lint/format fix

* feat: Can now set individual worker type names in bridgeData

* fix: wrong chec

* fix: avoid showing full debug payload

* chore: version bump hordelib

* chore: version bump hordelib

* branded_model

* feat: added --hordelib arg to update-runtime

* Create a Dockerfile for building horde worker with defaults, added docker instructions to README, github actions script for manually deploying docker image to GitHub container repo, generate_docker_bridge_config.py to generate a yaml for docker run

* merge

* no changelog

* feat: add --scribe

* feat: adds feature to save kudos model training data

* fix: don't assume post_processing exists in payload

* fix: elif

* fix: adds long prompt into kudos training data

* chore: bump hordelib version

* fix: remove obsolete GPUInfo

* feat: tweaks to worker debug

* fix: don't enable kudos calculation always

* fix: add missing kudos calculator import

* feat: simplify kudos model

* feat: simulate 0.5% kudos bonus per model

* fix: simulate kudos even when not saving training data

* fix: comment out a bunch of noisey debugs

* fix: make jobs per hour calculation more accurate

* chore: version bump hordelib

* feat: update kudos model v20.66

* feat: activate kudos simulation

* feat: remove per model bonus from simulation

* style: Pre-release format catch-up (#186)

* style: format catch-up

* style: remove unused code

* v22

* chore: bump hordelib version

* chore: bump hordelib version

* fix: change the display of free ram in the terminal ui

* fix: annoying UX at startup in terminal ui

* fix: stop simulating kudos locally

* chore: bump hordelib

* fix: terminal ui is now not case sensitive on key presses

---------

+++ update-runtime.cmd
@@ -12,5 +12,27 @@ IF EXIST CONDA GOTO WORKAROUND_END
 umamba create --no-shortcuts -r conda -n windows -f environment.yaml -y
 :WORKAROUND_END
 umamba create --no-shortcuts -r conda -n windows -f environment.yaml -y
-umamba run -r conda -n windows python -s -m pip install -r requirements.txt
-echo If there are no errors above everything should be correctly installed (If not, try running update_runtime.cmd as admin).
+
+set "hordelib="
+set "scribe="
+setlocal EnableDelayedExpansion
+for %%a in (%*) do (
+    if /I "%%a"=="--hordelib" set "hordelib=true"
+    if /I "%%a"=="--scribe" set    
+)
+endlocal
+
+REM Check if hordelib argument is defined
+if defined hordelib (
+  umamba run -r conda -n windows python -s -m pip uninstall -y hordelib
+  umamba run -r conda -n windows python -s -m pip install hordelib
+) else (
+  if defined scribe (
+    umamba run -r conda -n windows python -s -m pip install -r requirements-scribe.txt
+  ) else (
+    umamba run -r conda -n windows python -s -m pip uninstall nataili
+    umamba run -r conda -n windows python -s -m pip install -r requirements.txt
+  )
+)
+
+echo If there are no errors above everything should be correctly installed (If not, try deleting the folder /conda/envs/ and try again).
